+++ src/LogixDb.Data.Sqlite/Scripts/Migrations/Logix_202602131330_CreateTagConsumerTable.sql
@@ -4,7 +4,7 @@ CREATE TABLE tag_consumer
     producer        TEXT    NOT NULL,
     remote_tag      TEXT    NOT NULL,
     remote_instance INTEGER NOT NULL,
-    rpi             REAL    NOT NULL,
+    rpi             NUMERIC NOT NULL,
     unicast         INTEGER NOT NULL,
     record_hash     TEXT    NOT NULL,
     FOREIGN KEY (tag_id) REFERENCES tag (tag_id) ON DELETE CASCADE
